make base block and refactor blocks accordingly
